task_1_1_3: tests implemented

+++ Task_1_1_3/src/main/java/ru/nsu/svirsky/Add.java
@@ -10,7 +10,7 @@ public class Add extends Expression {
     }
 
     @Override
-    public float eval(String vars) {
+    public double eval(String vars) {
         return firstSummand.eval(vars) + secondSummand.eval(vars);
     }
